cheri: Apply the freepage.cocci semantic patch
Calls to free_page and free_pages that cast its (pointer)
argument to unsigned long must cast it to uintptr_t instead.

Apply like this:

	make COCCI=scripts/coccinelle/cheri/free_page.cocci \
		MODE=patch coccicheck

+++ arch/alpha/kernel/pci_iommu.c
@@ -426,7 +426,7 @@ try_again:
 
 	*dma_addrp = pci_map_single_1(pdev, virt_to_phys(cpu_addr), size, 0);
 	if (*dma_addrp == DMA_MAPPING_ERROR) {
-		free_pages((unsigned long)cpu_addr, order);
+		free_pages((uintptr_t)cpu_addr, order);
 		if (alpha_mv.mv_pci_tbi || (gfp & GFP_DMA))
 			return NULL;
 		/* The address doesn't fit required mask and we
@@ -453,7 +453,7 @@ static void alpha_pci_free_coherent(struct device *dev, size_t size,
 {
 	struct pci_dev *pdev = alpha_gendev_to_pci(dev);
 	dma_unmap_single(&pdev->dev, dma_addr, size, DMA_BIDIRECTIONAL);
-	free_pages((unsigned long)cpu_addr, get_order(size));
+	free_pages((uintptr_t)cpu_addr, get_order(size));
 
 	DBGA2("pci_free_consistent: [%llx,%zx] from %ps\n",
 	      dma_addr, size, __builtin_return_address(0));
