+++ arch/alpha/kernel/srm_env.c
@@ -77,7 +77,7 @@ static int srm_env_proc_show(struct seq_file *m, void *v)
 		ret = 0;
 	} else
 		ret = -EFAULT;
-	free_page((unsigned long)page);
+	free_page((uintptr_t)page);
 	return ret;
 }
 
@@ -115,7 +115,7 @@ static ssize_t srm_env_proc_write(struct file *file, const char __user *buffer,
 	}
 
  out:
-	free_page((unsigned long)buf);
+	free_page((uintptr_t)buf);
 	return res;
 }
