+++ arch/arc/kernel/setup.c
@@ -604,7 +604,7 @@ static int show_cpuinfo(struct seq_file *m, void *v)
 
 	seq_printf(m, arc_platform_smp_cpuinfo());
 
-	free_page((unsigned long)str);
+	free_page((uintptr_t)str);
 done:
 	seq_printf(m, "\n");
