+++ arch/arm64/kvm/nested.c
@@ -112,7 +112,7 @@ int kvm_vcpu_init_nested(struct kvm_vcpu *vcpu)
 		for (int i = kvm->arch.nested_mmus_size; i < num_mmus; i++)
 			kvm_free_stage2_pgd(&kvm->arch.nested_mmus[i]);
 
-		free_page((unsigned long)vcpu->arch.ctxt.vncr_array);
+		free_page((uintptr_t)vcpu->arch.ctxt.vncr_array);
 		vcpu->arch.ctxt.vncr_array = NULL;
 
 		return ret;
