+++ arch/powerpc/sysdev/xive/spapr.c
@@ -558,7 +558,7 @@ static void xive_spapr_cleanup_queue(unsigned int cpu, struct xive_cpu *xc,
 	alloc_order = xive_alloc_order(xive_queue_shift);
 	if (is_secure_guest())
 		uv_unshare_page(PHYS_PFN(__pa(q->qpage)), 1 << alloc_order);
-	free_pages((unsigned long)q->qpage, alloc_order);
+	free_pages((uintptr_t)q->qpage, alloc_order);
 	q->qpage = NULL;
 }
