+++ arch/riscv/kvm/mmu.c
@@ -598,7 +598,8 @@ void kvm_riscv_mmu_free_pgd(struct kvm *kvm)
 	spin_unlock(&kvm->mmu_lock);
 
 	if (pgd)
-		free_pages((unsigned long)pgd, get_order(kvm_riscv_gstage_pgd_size));
+		free_pages((uintptr_t)pgd,
+		           get_order(kvm_riscv_gstage_pgd_size));
 }
 
 void kvm_riscv_mmu_update_hgatp(struct kvm_vcpu *vcpu)
