+++ arch/s390/hypfs/hypfs_diag_fs.c
@@ -352,7 +352,7 @@ static int diag224_get_name_table(void)
 	if (!diag224_cpu_names)
 		return -ENOMEM;
 	if (diag224(diag224_cpu_names)) {
-		free_page((unsigned long)diag224_cpu_names);
+		free_page((uintptr_t)diag224_cpu_names);
 		return -EOPNOTSUPP;
 	}
 	EBCASC(diag224_cpu_names + 16, (*diag224_cpu_names + 1) * 16);
@@ -361,7 +361,7 @@ static int diag224_get_name_table(void)
 
 static void diag224_delete_name_table(void)
 {
-	free_page((unsigned long)diag224_cpu_names);
+	free_page((uintptr_t)diag224_cpu_names);
 }
 
 int __init __hypfs_diag_fs_init(void)
