+++ arch/s390/hypfs/hypfs_sprp.c
@@ -41,7 +41,7 @@ static unsigned long hypfs_sprp_diag304(void *data, unsigned long cmd)
 
 static void hypfs_sprp_free(const void *data)
 {
-	free_page((unsigned long) data);
+	free_page((uintptr_t) data);
 }
 
 static int hypfs_sprp_create(void **data_ptr, void **free_ptr, size_t *size)
@@ -56,7 +56,7 @@ static int hypfs_sprp_create(void **data_ptr, void **free_ptr, size_t *size)
 	if (rc != 1) {
 		*data_ptr = *free_ptr = NULL;
 		*size = 0;
-		free_page((unsigned long) data);
+		free_page((uintptr_t) data);
 		return -EIO;
 	}
 	*data_ptr = *free_ptr = data;
@@ -104,7 +104,7 @@ static int __hypfs_sprp_ioctl(void __user *user_area)
 	rc = copy_to_user(user_area, diag304, sizeof(*diag304)) ? -EFAULT : 0;
 out:
 	kfree(diag304);
-	free_page((unsigned long) data);
+	free_page((uintptr_t) data);
 	return rc;
 }
