+++ arch/s390/kernel/time.c
@@ -422,7 +422,7 @@ static void __init stp_reset(void)
 	if (rc == 0)
 		set_bit(CLOCK_SYNC_HAS_STP, &clock_sync_flags);
 	else if (stp_online) {
-		free_page((unsigned long) stp_page);
+		free_page((uintptr_t) stp_page);
 		stp_page = NULL;
 		stp_online = false;
 	}
