+++ arch/um/drivers/mconsole_kern.c
@@ -337,7 +337,7 @@ static int mem_config(char *str, char **error_out)
 				unplug_index = UNPLUGGED_PER_PAGE;
 			}
 
-			free_page((unsigned long) addr);
+			free_page((uintptr_t) addr);
 			unplugged_pages_count--;
 		}
 		else {
